requirements.txt
aiosqlite
discord.py
pyaml
python-dotenv
dnspython
PyNaCl
pillow
decorator
anyio
git+https://github.com/Rapptz/discord-ext-menus